requirements.txt
streamlit
pandas
duckdb
pyarrow
snowflake-connector-python[pandas]
snowflake-snowpark-python